+++ src/main/java/com/sparta/goodbite/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.sparta.goodbite.domain.waiting.config;
+package com.sparta.goodbite.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
