+++ src/main/java/com/sparta/goodbite/domain/auth/UserRole.java
@@ -1,4 +1,4 @@
-package com.sparta.goodbite.auth;
+package com.sparta.goodbite.domain.auth;
 
 import lombok.Getter;
