+++ src/main/java/com/sparta/goodbite/domain/auth/controller/AuthController.java
@@ -1,8 +1,8 @@
-package com.sparta.goodbite.auth.controller;
+package com.sparta.goodbite.domain.auth.controller;
 
-import com.sparta.goodbite.auth.service.AuthService;
+import com.sparta.goodbite.domain.auth.service.AuthService;
 import com.sparta.goodbite.common.response.MessageResponseDto;
-import com.sparta.goodbite.common.response.ResponseUtil;
+import com.sparta.goodbite.common.util.ResponseUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
