+++ src/main/java/com/sparta/goodbite/domain/auth/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.goodbite.auth.dto;
+package com.sparta.goodbite.domain.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
