+++ src/main/java/com/sparta/goodbite/exception/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.sparta.goodbite.exception;
 
 import com.sparta.goodbite.common.response.MessageResponseDto;
-import com.sparta.goodbite.common.response.ResponseUtil;
+import com.sparta.goodbite.common.util.ResponseUtil;
 import com.sparta.goodbite.exception.customer.CustomerException;
 import com.sparta.goodbite.exception.auth.AuthException;
 import com.sparta.goodbite.exception.menu.MenuException;
@@ -67,7 +67,7 @@ public class GlobalExceptionHandler {
         log.error("에러 발생: ", e);
         return ResponseUtil.of(e.getOperatingHourErrorCode().getHttpStatus(), e.getMessage());
     }
-  
+
     @ExceptionHandler(WaitingException.class)
     public ResponseEntity<MessageResponseDto> handleWaitingException(WaitingException e) {
         log.error("에러 발생: ", e);
